+++ swov1/src/funcs/dboGetPluginConfig.ts
@@ -36,8 +36,10 @@ export function dboGetPluginConfig(
 ): APIPromise<
   Result<
     components.DboDatabasePluginConfigResponse,
-    | errors.GetPluginConfigBadRequestError
-    | errors.GetPluginConfigNotFoundError
+    | errors.CommonBadRequestErrorResponse
+    | errors.CommonUnauthorizedErrorResponse
+    | errors.CommonNotFoundErrorResponse
+    | errors.CommonInternalErrorResponse
     | SwoError
     | ResponseValidationError
     | ConnectionError
@@ -63,8 +65,10 @@ async function $do(
   [
     Result<
       components.DboDatabasePluginConfigResponse,
-      | errors.GetPluginConfigBadRequestError
-      | errors.GetPluginConfigNotFoundError
+      | errors.CommonBadRequestErrorResponse
+      | errors.CommonUnauthorizedErrorResponse
+      | errors.CommonNotFoundErrorResponse
+      | errors.CommonInternalErrorResponse
       | SwoError
       | ResponseValidationError
       | ConnectionError
@@ -149,7 +153,7 @@ async function $do(
 
   const doResult = await client._do(req, {
     context,
-    errorCodes: ["400", "404", "4XX", "5XX"],
+    errorCodes: ["400", "401", "404", "4XX", "500", "5XX"],
     retryConfig: context.retryConfig,
     retryCodes: context.retryCodes,
   });
@@ -164,8 +168,10 @@ async function $do(
 
   const [result] = await M.match<
     components.DboDatabasePluginConfigResponse,
-    | errors.GetPluginConfigBadRequestError
-    | errors.GetPluginConfigNotFoundError
+    | errors.CommonBadRequestErrorResponse
+    | errors.CommonUnauthorizedErrorResponse
+    | errors.CommonNotFoundErrorResponse
+    | errors.CommonInternalErrorResponse
     | SwoError
     | ResponseValidationError
     | ConnectionError
@@ -176,8 +182,10 @@ async function $do(
     | SDKValidationError
   >(
     M.json(200, components.DboDatabasePluginConfigResponse$inboundSchema),
-    M.jsonErr(400, errors.GetPluginConfigBadRequestError$inboundSchema),
-    M.jsonErr(404, errors.GetPluginConfigNotFoundError$inboundSchema),
+    M.jsonErr(400, errors.CommonBadRequestErrorResponse$inboundSchema),
+    M.jsonErr(401, errors.CommonUnauthorizedErrorResponse$inboundSchema),
+    M.jsonErr(404, errors.CommonNotFoundErrorResponse$inboundSchema),
+    M.jsonErr(500, errors.CommonInternalErrorResponse$inboundSchema),
     M.fail("4XX"),
     M.fail("5XX"),
   )(response, req, { extraFields: responseFields });