+++ swov1/src/funcs/dboUpdateDatabase.ts
@@ -36,8 +36,10 @@ export function dboUpdateDatabase(
 ): APIPromise<
   Result<
     void,
-    | errors.UpdateDatabaseBadRequestError
-    | errors.UpdateDatabaseNotFoundError
+    | errors.CommonBadRequestErrorResponse
+    | errors.CommonUnauthorizedErrorResponse
+    | errors.CommonNotFoundErrorResponse
+    | errors.CommonInternalErrorResponse
     | SwoError
     | ResponseValidationError
     | ConnectionError
@@ -63,8 +65,10 @@ async function $do(
   [
     Result<
       void,
-      | errors.UpdateDatabaseBadRequestError
-      | errors.UpdateDatabaseNotFoundError
+      | errors.CommonBadRequestErrorResponse
+      | errors.CommonUnauthorizedErrorResponse
+      | errors.CommonNotFoundErrorResponse
+      | errors.CommonInternalErrorResponse
       | SwoError
       | ResponseValidationError
       | ConnectionError
@@ -150,7 +154,7 @@ async function $do(
 
   const doResult = await client._do(req, {
     context,
-    errorCodes: ["400", "404", "4XX", "5XX"],
+    errorCodes: ["400", "401", "404", "4XX", "500", "5XX"],
     retryConfig: context.retryConfig,
     retryCodes: context.retryCodes,
   });
@@ -165,8 +169,10 @@ async function $do(
 
   const [result] = await M.match<
     void,
-    | errors.UpdateDatabaseBadRequestError
-    | errors.UpdateDatabaseNotFoundError
+    | errors.CommonBadRequestErrorResponse
+    | errors.CommonUnauthorizedErrorResponse
+    | errors.CommonNotFoundErrorResponse
+    | errors.CommonInternalErrorResponse
     | SwoError
     | ResponseValidationError
     | ConnectionError
@@ -177,8 +183,10 @@ async function $do(
     | SDKValidationError
   >(
     M.nil(204, z.void()),
-    M.jsonErr(400, errors.UpdateDatabaseBadRequestError$inboundSchema),
-    M.jsonErr(404, errors.UpdateDatabaseNotFoundError$inboundSchema),
+    M.jsonErr(400, errors.CommonBadRequestErrorResponse$inboundSchema),
+    M.jsonErr(401, errors.CommonUnauthorizedErrorResponse$inboundSchema),
+    M.jsonErr(404, errors.CommonNotFoundErrorResponse$inboundSchema),
+    M.jsonErr(500, errors.CommonInternalErrorResponse$inboundSchema),
     M.fail("4XX"),
     M.fail("5XX"),
   )(response, req, { extraFields: responseFields });