+++ swov1/src/funcs/demListProbes.ts
@@ -32,7 +32,8 @@ export function demListProbes(
 ): APIPromise<
   Result<
     components.DemListProbesResponse,
-    | errors.ListProbesInternalServerError
+    | errors.CommonUnauthorizedErrorResponse
+    | errors.CommonInternalErrorResponse
     | SwoError
     | ResponseValidationError
     | ConnectionError
@@ -56,7 +57,8 @@ async function $do(
   [
     Result<
       components.DemListProbesResponse,
-      | errors.ListProbesInternalServerError
+      | errors.CommonUnauthorizedErrorResponse
+      | errors.CommonInternalErrorResponse
       | SwoError
       | ResponseValidationError
       | ConnectionError
@@ -120,7 +122,7 @@ async function $do(
 
   const doResult = await client._do(req, {
     context,
-    errorCodes: ["4XX", "500", "5XX"],
+    errorCodes: ["401", "4XX", "500", "5XX"],
     retryConfig: context.retryConfig,
     retryCodes: context.retryCodes,
   });
@@ -135,7 +137,8 @@ async function $do(
 
   const [result] = await M.match<
     components.DemListProbesResponse,
-    | errors.ListProbesInternalServerError
+    | errors.CommonUnauthorizedErrorResponse
+    | errors.CommonInternalErrorResponse
     | SwoError
     | ResponseValidationError
     | ConnectionError
@@ -146,7 +149,8 @@ async function $do(
     | SDKValidationError
   >(
     M.json(200, components.DemListProbesResponse$inboundSchema),
-    M.jsonErr(500, errors.ListProbesInternalServerError$inboundSchema),
+    M.jsonErr(401, errors.CommonUnauthorizedErrorResponse$inboundSchema),
+    M.jsonErr(500, errors.CommonInternalErrorResponse$inboundSchema),
     M.fail("4XX"),
     M.fail("5XX"),
   )(response, req, { extraFields: responseFields });