+++ swov1/src/funcs/demUnpauseUriMonitoring.ts
@@ -36,7 +36,9 @@ export function demUnpauseUriMonitoring(
 ): APIPromise<
   Result<
     components.CommonEntityId,
-    | errors.UnpauseUriMonitoringNotFoundError
+    | errors.CommonUnauthorizedErrorResponse
+    | errors.CommonNotFoundErrorResponse
+    | errors.CommonInternalErrorResponse
     | SwoError
     | ResponseValidationError
     | ConnectionError
@@ -62,7 +64,9 @@ async function $do(
   [
     Result<
       components.CommonEntityId,
-      | errors.UnpauseUriMonitoringNotFoundError
+      | errors.CommonUnauthorizedErrorResponse
+      | errors.CommonNotFoundErrorResponse
+      | errors.CommonInternalErrorResponse
       | SwoError
       | ResponseValidationError
       | ConnectionError
@@ -148,7 +152,7 @@ async function $do(
 
   const doResult = await client._do(req, {
     context,
-    errorCodes: ["404", "4XX", "5XX"],
+    errorCodes: ["401", "404", "4XX", "500", "5XX"],
     retryConfig: context.retryConfig,
     retryCodes: context.retryCodes,
   });
@@ -163,7 +167,9 @@ async function $do(
 
   const [result] = await M.match<
     components.CommonEntityId,
-    | errors.UnpauseUriMonitoringNotFoundError
+    | errors.CommonUnauthorizedErrorResponse
+    | errors.CommonNotFoundErrorResponse
+    | errors.CommonInternalErrorResponse
     | SwoError
     | ResponseValidationError
     | ConnectionError
@@ -174,7 +180,9 @@ async function $do(
     | SDKValidationError
   >(
     M.json(200, components.CommonEntityId$inboundSchema),
-    M.jsonErr(404, errors.UnpauseUriMonitoringNotFoundError$inboundSchema),
+    M.jsonErr(401, errors.CommonUnauthorizedErrorResponse$inboundSchema),
+    M.jsonErr(404, errors.CommonNotFoundErrorResponse$inboundSchema),
+    M.jsonErr(500, errors.CommonInternalErrorResponse$inboundSchema),
     M.fail("4XX"),
     M.fail("5XX"),
   )(response, req, { extraFields: responseFields });