+++ swov1/src/funcs/entitiesGetEntityById.ts
@@ -36,9 +36,10 @@ export function entitiesGetEntityById(
 ): APIPromise<
   Result<
     components.EntitiesEntity,
-    | errors.GetEntityByIdBadRequestError
-    | errors.GetEntityByIdUnauthorizedError
-    | errors.GetEntityByIdNotFoundError
+    | errors.CommonBadRequestErrorResponse
+    | errors.CommonUnauthorizedErrorResponse
+    | errors.CommonNotFoundErrorResponse
+    | errors.CommonInternalErrorResponse
     | SwoError
     | ResponseValidationError
     | ConnectionError
@@ -64,9 +65,10 @@ async function $do(
   [
     Result<
       components.EntitiesEntity,
-      | errors.GetEntityByIdBadRequestError
-      | errors.GetEntityByIdUnauthorizedError
-      | errors.GetEntityByIdNotFoundError
+      | errors.CommonBadRequestErrorResponse
+      | errors.CommonUnauthorizedErrorResponse
+      | errors.CommonNotFoundErrorResponse
+      | errors.CommonInternalErrorResponse
       | SwoError
       | ResponseValidationError
       | ConnectionError
@@ -149,7 +151,7 @@ async function $do(
 
   const doResult = await client._do(req, {
     context,
-    errorCodes: ["400", "401", "404", "4XX", "5XX"],
+    errorCodes: ["400", "401", "404", "4XX", "500", "5XX"],
     retryConfig: context.retryConfig,
     retryCodes: context.retryCodes,
   });
@@ -164,9 +166,10 @@ async function $do(
 
   const [result] = await M.match<
     components.EntitiesEntity,
-    | errors.GetEntityByIdBadRequestError
-    | errors.GetEntityByIdUnauthorizedError
-    | errors.GetEntityByIdNotFoundError
+    | errors.CommonBadRequestErrorResponse
+    | errors.CommonUnauthorizedErrorResponse
+    | errors.CommonNotFoundErrorResponse
+    | errors.CommonInternalErrorResponse
     | SwoError
     | ResponseValidationError
     | ConnectionError
@@ -177,9 +180,10 @@ async function $do(
     | SDKValidationError
   >(
     M.json(200, components.EntitiesEntity$inboundSchema),
-    M.jsonErr(400, errors.GetEntityByIdBadRequestError$inboundSchema),
-    M.jsonErr(401, errors.GetEntityByIdUnauthorizedError$inboundSchema),
-    M.jsonErr(404, errors.GetEntityByIdNotFoundError$inboundSchema),
+    M.jsonErr(400, errors.CommonBadRequestErrorResponse$inboundSchema),
+    M.jsonErr(401, errors.CommonUnauthorizedErrorResponse$inboundSchema),
+    M.jsonErr(404, errors.CommonNotFoundErrorResponse$inboundSchema),
+    M.jsonErr(500, errors.CommonInternalErrorResponse$inboundSchema),
     M.fail("4XX"),
     M.fail("5XX"),
   )(response, req, { extraFields: responseFields });