+++ swov1/src/funcs/entitiesListEntities.ts
@@ -44,8 +44,9 @@ export function entitiesListEntities(
   PageIterator<
     Result<
       operations.ListEntitiesResponse,
-      | errors.ListEntitiesBadRequestError
-      | errors.ListEntitiesUnauthorizedError
+      | errors.CommonBadRequestErrorResponse
+      | errors.CommonUnauthorizedErrorResponse
+      | errors.CommonInternalErrorResponse
       | SwoError
       | ResponseValidationError
       | ConnectionError
@@ -74,8 +75,9 @@ async function $do(
     PageIterator<
       Result<
         operations.ListEntitiesResponse,
-        | errors.ListEntitiesBadRequestError
-        | errors.ListEntitiesUnauthorizedError
+        | errors.CommonBadRequestErrorResponse
+        | errors.CommonUnauthorizedErrorResponse
+        | errors.CommonInternalErrorResponse
         | SwoError
         | ResponseValidationError
         | ConnectionError
@@ -165,7 +167,7 @@ async function $do(
 
   const doResult = await client._do(req, {
     context,
-    errorCodes: ["400", "401", "4XX", "5XX"],
+    errorCodes: ["400", "401", "4XX", "500", "5XX"],
     retryConfig: context.retryConfig,
     retryCodes: context.retryCodes,
   });
@@ -180,8 +182,9 @@ async function $do(
 
   const [result, raw] = await M.match<
     operations.ListEntitiesResponse,
-    | errors.ListEntitiesBadRequestError
-    | errors.ListEntitiesUnauthorizedError
+    | errors.CommonBadRequestErrorResponse
+    | errors.CommonUnauthorizedErrorResponse
+    | errors.CommonInternalErrorResponse
     | SwoError
     | ResponseValidationError
     | ConnectionError
@@ -194,8 +197,9 @@ async function $do(
     M.json(200, operations.ListEntitiesResponse$inboundSchema, {
       key: "Result",
     }),
-    M.jsonErr(400, errors.ListEntitiesBadRequestError$inboundSchema),
-    M.jsonErr(401, errors.ListEntitiesUnauthorizedError$inboundSchema),
+    M.jsonErr(400, errors.CommonBadRequestErrorResponse$inboundSchema),
+    M.jsonErr(401, errors.CommonUnauthorizedErrorResponse$inboundSchema),
+    M.jsonErr(500, errors.CommonInternalErrorResponse$inboundSchema),
     M.fail("4XX"),
     M.fail("5XX"),
   )(response, req, { extraFields: responseFields });
@@ -213,8 +217,9 @@ async function $do(
     next: Paginator<
       Result<
         operations.ListEntitiesResponse,
-        | errors.ListEntitiesBadRequestError
-        | errors.ListEntitiesUnauthorizedError
+        | errors.CommonBadRequestErrorResponse
+        | errors.CommonUnauthorizedErrorResponse
+        | errors.CommonInternalErrorResponse
         | SwoError
         | ResponseValidationError
         | ConnectionError