+++ swov1/src/funcs/metricsGetMetricByName.ts
@@ -47,7 +47,9 @@ export function metricsGetMetricByName(
   PageIterator<
     Result<
       operations.GetMetricByNameResponse,
-      | errors.GetMetricByNameNotFoundError
+      | errors.CommonUnauthorizedErrorResponse
+      | errors.CommonNotFoundErrorResponse
+      | errors.CommonInternalErrorResponse
       | SwoError
       | ResponseValidationError
       | ConnectionError
@@ -76,7 +78,9 @@ async function $do(
     PageIterator<
       Result<
         operations.GetMetricByNameResponse,
-        | errors.GetMetricByNameNotFoundError
+        | errors.CommonUnauthorizedErrorResponse
+        | errors.CommonNotFoundErrorResponse
+        | errors.CommonInternalErrorResponse
         | SwoError
         | ResponseValidationError
         | ConnectionError
@@ -162,7 +166,7 @@ async function $do(
 
   const doResult = await client._do(req, {
     context,
-    errorCodes: ["404", "4XX", "5XX"],
+    errorCodes: ["401", "404", "4XX", "500", "5XX"],
     retryConfig: context.retryConfig,
     retryCodes: context.retryCodes,
   });
@@ -177,7 +181,9 @@ async function $do(
 
   const [result, raw] = await M.match<
     operations.GetMetricByNameResponse,
-    | errors.GetMetricByNameNotFoundError
+    | errors.CommonUnauthorizedErrorResponse
+    | errors.CommonNotFoundErrorResponse
+    | errors.CommonInternalErrorResponse
     | SwoError
     | ResponseValidationError
     | ConnectionError
@@ -190,7 +196,9 @@ async function $do(
     M.json(200, operations.GetMetricByNameResponse$inboundSchema, {
       key: "Result",
     }),
-    M.jsonErr(404, errors.GetMetricByNameNotFoundError$inboundSchema),
+    M.jsonErr(401, errors.CommonUnauthorizedErrorResponse$inboundSchema),
+    M.jsonErr(404, errors.CommonNotFoundErrorResponse$inboundSchema),
+    M.jsonErr(500, errors.CommonInternalErrorResponse$inboundSchema),
     M.fail("4XX"),
     M.fail("5XX"),
   )(response, req, { extraFields: responseFields });
@@ -208,7 +216,9 @@ async function $do(
     next: Paginator<
       Result<
         operations.GetMetricByNameResponse,
-        | errors.GetMetricByNameNotFoundError
+        | errors.CommonUnauthorizedErrorResponse
+        | errors.CommonNotFoundErrorResponse
+        | errors.CommonInternalErrorResponse
         | SwoError
         | ResponseValidationError
         | ConnectionError