+++ swov1/src/funcs/metricsListMetricAttributes.ts
@@ -47,7 +47,9 @@ export function metricsListMetricAttributes(
   PageIterator<
     Result<
       operations.ListMetricAttributesResponse,
-      | errors.ListMetricAttributesNotFoundError
+      | errors.CommonUnauthorizedErrorResponse
+      | errors.CommonNotFoundErrorResponse
+      | errors.CommonInternalErrorResponse
       | SwoError
       | ResponseValidationError
       | ConnectionError
@@ -76,7 +78,9 @@ async function $do(
     PageIterator<
       Result<
         operations.ListMetricAttributesResponse,
-        | errors.ListMetricAttributesNotFoundError
+        | errors.CommonUnauthorizedErrorResponse
+        | errors.CommonNotFoundErrorResponse
+        | errors.CommonInternalErrorResponse
         | SwoError
         | ResponseValidationError
         | ConnectionError
@@ -173,7 +177,7 @@ async function $do(
 
   const doResult = await client._do(req, {
     context,
-    errorCodes: ["404", "4XX", "5XX"],
+    errorCodes: ["401", "404", "4XX", "500", "5XX"],
     retryConfig: context.retryConfig,
     retryCodes: context.retryCodes,
   });
@@ -188,7 +192,9 @@ async function $do(
 
   const [result, raw] = await M.match<
     operations.ListMetricAttributesResponse,
-    | errors.ListMetricAttributesNotFoundError
+    | errors.CommonUnauthorizedErrorResponse
+    | errors.CommonNotFoundErrorResponse
+    | errors.CommonInternalErrorResponse
     | SwoError
     | ResponseValidationError
     | ConnectionError
@@ -201,7 +207,9 @@ async function $do(
     M.json(200, operations.ListMetricAttributesResponse$inboundSchema, {
       key: "Result",
     }),
-    M.jsonErr(404, errors.ListMetricAttributesNotFoundError$inboundSchema),
+    M.jsonErr(401, errors.CommonUnauthorizedErrorResponse$inboundSchema),
+    M.jsonErr(404, errors.CommonNotFoundErrorResponse$inboundSchema),
+    M.jsonErr(500, errors.CommonInternalErrorResponse$inboundSchema),
     M.fail("4XX"),
     M.fail("5XX"),
   )(response, req, { extraFields: responseFields });
@@ -219,7 +227,9 @@ async function $do(
     next: Paginator<
       Result<
         operations.ListMetricAttributesResponse,
-        | errors.ListMetricAttributesNotFoundError
+        | errors.CommonUnauthorizedErrorResponse
+        | errors.CommonNotFoundErrorResponse
+        | errors.CommonInternalErrorResponse
         | SwoError
         | ResponseValidationError
         | ConnectionError